# src/CalculatorAPI/Controllers/UserApi.cs
/*
 * Calculator API
 *
 * Calculator API in .NET 7 
 *
 * OpenAPI spec version: 1.0.0
 * Contact: tesfai80@hotmail.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using CalculatorAPI.Attributes;
using CalculatorAPI.Interfaces;
using CalculatorAPI.Models;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using System;

namespace CalculatorAPI.Controllers
{
    /// <summary>
    /// User Controller 
    /// </summary>
    [ApiController]
    public class UserApiController : ControllerBase
    {
        private IUserService _userService;
        /// <summary>
        /// Injecting IuserIUserServiceService here
        /// </summary>
        /// <param name="service"></param>
        public UserApiController(IUserService userService )
        {
            _userService = userService;
        }
        /// <summary>
        /// Login User user and returns JWT token
        /// </summary>
        /// <param name="user"></param>
        /// <response code="200">Successful login</response>
        /// <response code="400">Bad request</response>
        /// <response code="401">Unauthorized</response>
        [HttpPost]
        [Route("/TESFAI80/CalculatorAPI/1.0.0/user")]
        [ValidateModelState]
        [SwaggerOperation("LoginUser")]
        public virtual IActionResult LoginUser([FromBody]User user)
        {
            var token = _userService.Login(user.UserName, user.Password);

            if (token == null || token == String.Empty)
                return BadRequest(new { message = "User name or password is incorrect" });

            return Ok(token);
        }
    }
}